Automatically convert untyped scratch allocators
C++ standard library containers require that their allocator classes be
template specialized to allocate a single type of object.  However, the
standard types that manage the allocation of scratch (per-compile)
memory do not specialize based on the type their allocating.  This in
turn requires that developers desiring to use one of those scratch
allocators explicitly instantiate a TR::typed_allocator with the
appropriate configuration of untyped delegate allocator.

To reduce boilerplate and [hopefully] ease readability, this commit
provides conversion operators for TR::Region and CS2::shared_allocator
(used as TR::Allocator) to allow them to automatically convert to the
appropriate form of TR::typed_allocator, and consequently allow them to
be passed directly as a parameter when constructing a standard library
container.

Signed-off-by: Luc des Trois Maisons <lmaisons@ca.ibm.com>