Fix disableLongDispStackSlot option on Z
This patch fixes two bugs when long displacement stack slot is disabled

First, stop ExtCodeBaseRegister GPR7/HPR7 from being incorrectly added
to the Global Register Table when it is not free for assignment.

Second, fix LG, LGF, and LLGF binary encoding to use the right source
and scratch registers when the ExtCodeBaseRegister is available.

Signed-off-by: John Xu <xujohn@ca.ibm.com>