Add IL Verifiers
In this case, verifiers are run after optimizations take place.
However, the idea of verifiers are generic - it is possible to move
current IL verification, such as `verify{Trees,Blocks,CFG}` to IL
verifiers, and out of `TR_Debug`.

If simple hooks are added to `OMRCompilation`, this infrastructure
can also be used by downstream projects to choose how and when IL
should be verified, without changing OMR.

Signed-off-by: Lukas Lorimer <noryb009@gmail.com>