Add optimization test driver
An optimization test driver is meant to make it easy to set up a
test for an optimization. It allows you to add an IlInjector,
a sequence of optimizations, and a verifier. The IL is then
generated, optimized and verified.

From that point, codegen is optional. If `Verify` is called,
the compilation will stop there. If `VerifyAndInvoke` is called
and the IL verification succeeded, the tests in `invokeTests`
will be run.

Signed-off-by: Lukas Lorimer <noryb009@gmail.com>