Support for aborted Concurrent Scavenger during root processing
If aborted scavenge occurs during root processing (rare, but still
possible if a very large object is directly referenced from the roots)
proceed with the rest of the steps in the cycle while still in STW.
Currently, we would resume the app and concurrent scavenge operation as
if nothing happened, which is dangerous since not all roots might not
properly updated.
The code changes are contained within CS specific state machine.

We will now be able to report all the necessary events, and it will be
as clear in verbose gc that aborted cycle occurred, is if it were the
more common case (during main scan phase).

Also, piggybacking some general (not only CS specific) cleanup in
Scavenger - SemiSpace interaction relevant to the cycle
setup/teardown/abort.

Signed-off-by: Aleksandar Micic <amicic@ca.ibm.com>