Fix restoreTOC
It currently broke Linux32. Also, it is wrong on multi-fronts.

Obviously JIT pseudo-TOC is mixed up with JIT-module system TOC. Furthermore, currently it restored TOC blindly and led to at least two issues:
a) Linux32 doesn't have system TOC, and it crashed in compilation;
b) Restore pseudo-TOC unnecessarily, and potentially have load-hit-store performance hazard;

As the name implies, JIT pseudo-TOC is used by JIT-ed codes, thus there is nothing to do with SystemLinkage. We should forget about JIT pseudo-TOC completely, when we dealt with SystemLinkage. In AIX-style systemLinkage, there is a system TOC in each module (shared library or executable, that is) whose base-address needs to be set up in gr2 by the caller before the call (LE system, callee sets up itself). That base-address is carried in the function descriptor which contains the function address too. And, gr2 needs to be restored to the caller's module TOC's base-address after the call returns. However, JIT-ed code doesn't use system TOC at all (JIT-ed code is not a module either), such that we don't need to restore TOC base-address (as a matter of fact, gr2 in JIT-ed code is a scratch register anyhow). For static compiler, on the other hand, you do need to generate the gr2 restoration.

Since this codebase is not intended to generate ABI-conforming module of dynamic code, I commented out the gr2 restoration completely (another way: you can restore it for only aix-style, excluding Linux32, that is).

Fixed issue: https://github.com/eclipse/omr/issues/611
Signed-off-by: Julian Wang <zlwang@ca.ibm.com>