Fix to return correct currentMethod/Symbol for comp
Before this fix, *getCurrentMethod* and *getMethodSymbol* API on comp
return compiling(caller) method/symbol for ilgen but callee method/symbol for
optimization. The reason behind this is that ilgen needs the 'caller' in
some cases and *_methodSymbol* on *IlGenerator* would keep callee
method/symbol if 'callee' is needed. However, this can cause
inconsistency between the ilgen and optimizer when a temp is created
in ilgen but analyzed in optimizer.

This fix changes the 2 APIs to make ilgen and optimizer get the same
current method or symbol. The compiling method is indeed needed in
several places and there is another API getJittedMethodSymbol on comp
for this.

The changes are in 3 kinds:
1. add methodSymbol() to TR_IlGenerator to facilitate 'kind 2'
2. change getCurrentMethod and getMethodSymbol on comp to return callee
symbol
3. fix several places should have used getJittedMethodSymbol rather than
getMethodSymbol

Signed-off-by: Yi Zhang <yizhang@ca.ibm.com>