Fix checkSize, extend grow function in operand stack
checkSize() was not updated to test the correct condition for
_stackTop when it was recently modified.

grow() does not currently take a hint amount, and is also not
correct for very small stack sizes. Added a hint amount which
defaults to 0. When 0 is passed in, grow() will try to grow
the stack by _stackMax >> 1 (i.e. grow by 50%). If the stack
only has a single element, even this grown amount will be 0,
so bump it to 1 in this case.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>