Implement UnionType class in JitBuilder
The `UnionType` class can be used internally by JitBuilder to represent
union types. In it's current state, it is heavily based on the
`StructType` class.

This commit addresses the problem of representing and manipulating
unions *within* JitBuilder. Most importantly, it provides a way for
JitBuilder to define the fields of a union and alias them together.
However, it does not add an interface to JitBuilder for defining
and actually using unions. This will done in a later commit.

Issue: #643

Signed-off-by: Leonardo Banderali <leob@linux.vnet.ibm.com>