Reinstate z9 / z10 ALS
Reinstate z9 / z10 architecture level set (ALS) in the OMR z Systems
codegen. This is the minimal architecture set which is to be supported
by compiler built on top of OMR moving forward. Hence we need to modify
all locations that generate instructions introduced in subsequent
iterations of  z/Architecture.

This change effectively guards many sequences that we generate in the
compiler with architecture checks, and provides alternative
implementations that are supported on z9 / z10 if needed.

Issue: #667
Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>