Move ineliminable sign-extensions in loop strider
When loop strider replaces a 32-bit induction variable `i` with a new
64-bit `k` within a loop, the original loads of `i` in that loop
become `l2i` of loads of `k`. Such an `l2i` node may occur under
arithmetic in a context such as the following, where it is not
possible to eliminate the outer `i2l` node.

    i2l
      iadd (cannotOverflow)
        iload foo
        l2i
          lload k

In this case, changing the `i2l` into `ladd` would require creating a
new `i2l` wrapping `iload foo`, which so far strider elects never to
do. But with this change, the above tree will now transform into the
following tree that does fewer conversions:

    ladd (cannotOverflow)
      i2l
        iload foo
      lload k

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>