Issue with pseudo-TOC mulitplexing on sys-TOC reg
Unforfunately, Ruby JIT multiplexed gr2 for JIT pseudo-TOC.
I figured it is true as well for Python and jittest.

Since gr2 doesn't participate in sysLinkage reg assigning and
there is only one pseudo-TOC in existence, I don't restore
that not-changed pseudo-TOC on the way out for the multiplexing
case for performance reason of course.

Tested against both Ruby and Java.

Fixed issue: #665

Signed-off-by: Julian Wang <zlwang@ca.ibm.com>