Flag for OSR infrastructure removal
OSR infrastructure must be present to support induce OSR, however,
components of this infrastructure, particularly exception edges, may
have been removed. This adds a flag to determine if the exception
edges have been removed and extends existing functions to consider it.

Also modifies conditions for induce OSR, such that it is supported
along side MimicInterpreterFrameShape in the event of full speed
debugging.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>