Add OSR points after monents
This extends OSR to add points after a monent, similar to what is done
with calls and asyncchecks. These points are used for analysis and
may result in the insertion of OSR guards. This breaks the assumption
that OSR points are always before and optionally after a node. The
code has been altered to make them both optional, but one mandatory.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>