Fix bug in rematerialization safety logic
The RematTools serve to allow you to try and recompute a store on a
different control flow path. This is used in the inliner to do
privatized inliner argument rematerialization in the blocks on the taken
side of virtual guards. The RematTools include a bunch of safety
checking logic to ensure it is safe to remat a store. There is a bug in
this remat logic that could cause some nodes that should be checked to
be skipped. This patch fixes this problem and adds additional helpful
trace to the RematTools.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>