Validated calls to getHighGlobalRegisterNumber
On 64 bit machines, the getHighGlobalRegisterNumber call is not valid
nor needed. On 32 bit machines, a register dependency node's children
must be checked to see if they are of the long type, before performing
the getHighGlobalRegisterNumber call.

If the register dependency node has a child without a type, its child,
and every subsequent child must be checked until a descendant is found
with a type. If this descendant node is of the long type, then the
getHighGlobalRegisterNumber call will be performed on the original node,
which would be the register dependency node's first child.

Therefore the following criteria need to be valid to perform
the call on a register dependency node's child:

- The target machine must be 32 bit
- A descendant must have the long type

A helper function was created to call getHighGlobalRegisterNumber only
after confirming the above mentioned criteria. Otherwise the function
will return a -1.

Signed-off-by: Ryan Santhirarajan <rsanth@ca.ibm.com>