Provide default for getClassFromNewArrayType
Returning NULL is an acceptable choice, as it is always possible that
the class for a newarray cannot be determined.

This particular change is driven by Ruby+OMR attempting to use VP, when
Ruby+OMR generates a local array and loads its address using
TR::loadaddr. VP wants to add a constraint based on the type of the
array, but cannot because the array isn't object typed.

Signed-off-by: Matthew Gaudet <magaudet@ca.ibm.com>