Vector Eval bug fixes and perf improvements
Added more entries to convertScalarToVector such as frem, drem and lxor
which are converted to vrem, vrem and vxor.

Changed vsplats evaluator on Power to use direct move instructions to
transfer data from GPRs to vector registers instead of transferring the
data through memory. This applies to splatting Int32 and Int64 data
types since Double and Float did not transfer data through memory.

Fixed a bug in the vsplats evaluator on Power for Float data where a
TR_VSX_VECTOR register was used for the result register. It was changed
to a TR_VRF register because float data may use VMX instructions in
other places.

generateMvFprGprInstructions normally generates direct move instructions
to move data between GPRs and FPRs on Power. It has a fallback path
which transfers the data through memory instead when direct move isn't
supported. The fallback path was modified to support little endian
power.

getSupportsOpCodeForAutoSIMD on Power and z were modified to more
accurately represent which vector evaluators are currently supported and
for which data types.

vand, vor and vxor evaluators on Power were expanded to support Int64
data type. Int64 vectors uses the VSX registers so the previously used
VMX instruction was invalid for Int64 data.

Changed getvelem evaluator on Power to use direct move instructions to
transfer data from vector registers to GPRs instead of transferring the
data through memory.

Fixed bug with vneg evaluator on Power. Using the vsubsws would produce
the wrong result when the input was -2147483648 (minimum signed 32 bit
int) due to being a saturate instruction. It was changed to vsubuwm
which is a modulo instruction.

Fixed a bug with getvelem evaluator on z. When operating on 64 bit data
on a 31 bit system, the result is supposed to be returned in a pair
register. This was not happening and has now been corrected.

Fixed a bug with vsetelem evaluator on z. When operating on 64 bit data
on a 31 bit system, the wrong parts of the vector register were being
read out into the pair register due to the index calculation being
wrong. This was fixed.

Issue: #767
Signed-off-by: jimmyk <jimmyk@ca.ibm.com>