Provide managed destruction of Region allocations
Under some circumstances, we will want to Region-allocate objects that
require having their destructors invoked when they are destroyed.
Currently there is no mechanism for doing this.  This change set
provides such a mechanism for types that would allow themselves to be
copy-constructed into the Region.

In order to provide this new facility, the allocate function is extended
via a new function template such that invocations passing a parameter
that is not coercable to size_t are treated as instead providing an
object prototype with which to construct a new instance via copying.
When the Region is itself destroyed, it will iterate through these
managed objects, destroying them in LIFO ordering.

Signed-off-by: Luc des Trois Maisons <lmaisons@ca.ibm.com>