Add exception asynccheck insertion opt
Adds the exception handler asynccheck insertion optimization to the
ilgen opt strategy and a tracing option.

Corrects the case where an asynccheck has been inserted into an
exception handler, which is being split to use a goto. In this
situation, the original asynccheck is retained in the exception
handler and a duplicate is added to the new block. This prevents
the goto edge from creating a loop without an asynccheck and
ensures the exception handler still has its asynccheck.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>