Update append instruction when moving or removing
When an instructon is moved or removed we must be careful to properly
update the current append instruction pointer to which other
instructions will be attached to in the instruction linked list. Failing
to do so can have functional problems such as instructions not being
generated into the instruction stream where one would expect them.

In addition to this fix the Instruction::move API can reuse
Instruction::remove API to eliminate some code duplication.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>