Value Propagation performance bug fix
A feature that avoids doing VP on low frequency blocks (in order
to save compilation time) has a bug such that the JIT will not
do VP at hot/scorhing opt levels for blocks that are NOT cold.
The condition should be reversed so that the JIT avoids VP for
basic blocks that ARE cold.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>