Inhibit array splitting in scavenger backout
This was previously done for the RS overflow case but also needs to
be done in non-RS overflow backout. This fix covers both cases. The
boolean backOutFlag member has been extended to 3 states to distinguish
backOutFlagsRaised and backOutStarted conditions. Array splitting is 
inhibited only in backOutStarted state.

This additional change is necessary to limit inhibition of array 
splitting to state where backout thread has synchronized with other 
collecting threads and is proceeding with completeBackOut(). This allows
threads that are still actively scanning and haven't yet recognized
backout condition to split current array and complete split segment 
in shorter time and reach backout sync point sooner.

Signed-off-by: Kim Briggs <briggs@ca.ibm.com>