Improve IVA's usesUnchangedValueInLoopTest analysis
In order to identify an induction variable (IV) as "primary," induction
variable analysis (IVA) must determine, for the loop exit test against
that IV, which value is tested: the updated value, or the original from
the beginning of the iteration. If IVA fails to make this determination,
it must reject the candidate.

When the value tested is not directly from a load, IVA requires it to be
the right-hand side of the most recent store to the candidate IV.

On the other hand, when the value *is* directly from a load, IVA looks
at the stores to the same IV earlier in the extended block, in order to
determine whether the load is evaluated before them, or afterward. So
far this analysis has required the tested load to appear beneath each
such store, i.e. "within" the right-hand side, where it would need to be
evaluated before storing. With this requirement, IVA conservatively
rejects a primary IV given the following IV update and loop test:

        treetop
    n0n   iload i
        istore i
          iadd
    n1n     iload i
            iconst -1
        ificmple --> exit
    n0n   ==>iload i
          iconst 0

This commit changes the analysis to determine the relative order of
evaluation of the load and stores of interest by iterating over the
nodes in the extended block in postorder instead.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>