Expanded support of bitwise vector Evals on p
My previous commit would cause an assert to be thrown if an unexpected
data type is used with a bitwise vector opcode like vxor. I decided this
was not necessary since bitwise operations can be done regardless of the
actual type of the data. If an unexpected type is encountered, the VSX
bitwise instruction will be used since it works for any of the vector
registers on p.

Issue: #767
Signed-off-by: jimmyk <jimmyk@ca.ibm.com>