OMR codegen full support for certain evaluators
integerPairMulEvaluator
integerPairShlEvaluator
integerPairRolEvaluator
integerPairShrEvaluator
integerPairUshrEvaluator

For integerPairMulEvaluator, it does not handle the case
with non-constant multiplicand when TR_NoInlineLongMultiply
is set. The env var is purely for backing off to deprecated
code. Hence the path is removed.

For the shift evaluators, the whole code paths handling
non-constant shift amount are missing. Therefore, this commit
implements them.

Signed-off-by: Victor Ding <dvictor@ca.ibm.com>