Use resolved method for skipping HCR decisions
Callee method symbols may be null depending on whether we are peeking
inside of a top-level callee method to be inlined or not. For a method
two call stacks below the current compiled body the callee method symbol
may be null as it may not have been constructed yet. This means we are
unable to properly make a decision on whether or not to generate HCR
guards for such methods.

Alternatively a resolved method object should always be available for an
inline callee candidate. Thus we can always use the resolved method
object for determining whether or not to skip generating HCR guards.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>