Add efficient single bit dataflow analysis
The current dataflow engine is instantiated using BitVectors as backing
storage to support analysis requiring an arbitrary number of bits. While
flexible, the use of BitVectors is especially inefficient for single-bit
data flow analyses. This change introduces a new TR_SingleBitContainer
which implements the BitVector API used by the dataflow engine, but
which stores only a single boolean value. The dataflow framework is
instantiated for this new container to allow for much cheaper single-bit
dataflow analyses to be written.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>