Restore id field on CFGEdge
It can be very handy, during optimization, to be able to identify CFG
edges by an ID number. The OMR code has all the required APIs, but the
backing field is not present. This change restores the backing field so
that an optimization can store an ID number of a CFGEdge to simplify
booking keeping and analysis data structures. There is no requirement
for IDs to be unique or to have any meaning outside ofa single
optimization - it is simply a convenience field.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>