Add StoreOver service to IlBuilder
In some scenarios, it's useful to overwrite the value of the local
variable holding one IlValue with the value of another one. The
StoreOver service takes one value and overwrites it into the same
local variable of another IlValue (dest).

It's used in VirtualMachineOperandStack in the MergeInto operation.
In this case, other operations may have assumed their inputs are
in specific IlValues, so when MergeInto is used to create a second
control flow path to a particular IlBuilder, the IlValues in
operand stack slots *must* be written into the same local variables
used by the IlValues in corresponding locations in the operand
stack for the first control flow path. StoreOver basically reaches
into an IlValue to extract the name of the local variable used
to hold its value and then calls Store to actually write the
new value into that local variable.

Now that StoreOver exists, VirtualMachineOperandStack (and any
other consumers) no longer need to know how to reach into an
IlValue to extract its name.

Because VirtualMachineOperandStack was changed, the operandstacktests
test should already exercise the new service.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>