Code cleanup for bitwise vector Evals on p
Removed VectorInt64 case from the switch statments inside the evaluators
for vand, vor and vxor. The was done because handling the VectorInt64 case
was exactly the same as the default case so VectorInt64 did not need its
own case. This will slightly reduce the code size of the three vector
evaluators.

Issue: #767
Signed-off-by: jimmyk <jimmyk@ca.ibm.com>