Fix printf format string warnings in JitBuilder samples
Three of the samples have printf's passing int64_t into a printf
argument whose format descriptor is either %ld or %lx. These format
descriptors should be %lld and %llx, respectively.

With this change, all the JitBuilder samples compile without
any warnings and it would be nice to keep it that way going forward.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>