Track all CP symbols using cpSymRefs
    - rename cpSymRefs into cpConstantSymRefs since this bitvector currently tracks only constants in CP
    - add new cpSymRefs bitvector for tracking all symbol references in CP
    - set cpSymRefs whenever cpConstantSymRefs is set

cpSymRefs can now be used for tracking any symbol in CP. These symbols will be automatically dis-aliased from unsafe array shadows since there is already code for that in Aliases.cpp. The only difference is that this used to be done only for constants and now all CP symbols will be dis-aliased from unsafe array shadows, as long as they are included into cpSymRefs by a particular language implementation.

Signed-off-by: Gita Koblents <koblents@ca.ibm.com>