Correct off by 1 mistake in test for real method call
Whilst calls are potential OSR points, helper calls and
common non-helper symbols are not. This distinction is tested by
comparing the call's symbol reference number with the index of the
last common non-helper symbol. This index marks the end of the
predefined symbols, but can be used as the reference number for a
real method call.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>