Add Unsigned comparisons to JitBuilder
Unsigned compares are really useful, but currently not provided
by JitBuilder in IlBuilder or BytecodeBuilder. The OMR compiler
supports unsigned comparisons, so there's no reason they shouldn't
be available to JitBuilder.

I added Unsigned varieties for all comparisons in IlBuilder and
BytecodeBuilder. In the process, I managed to simplify some of
the logic used to generate conditional operations by consolidating
on TR_ComparisonType as the mechanism to get the appropriate
comparison opcode based on the datatype and whether the compare
should be signed/unsigned. I also modified integerizeAddresses()
to operate independently on the two operands provided to it.
This function is a workaround for several opcodes with Address
datatype (e.g. ifacmplt) whose evaluators are not implemented
for all platforms.

Issue: #833
Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>