Add IfAnd and IfOr to JitBuilder
To make it easier to construct nests of short-circuited conditionals,
I created IfAnd and IfOr. The process becomes smoother, at least.

The first two parameters for IfAnd and IfOr are IlBuilder objects
representing the end results of the AND / OR (respectively) nest.
If all conditional terms are true, then IfAnd will cause the operations
in "allTrueBuilder" to execute. If not, the "anyFalseBuilder"
operations will be executed.  Similarly for IfOr: "anyTrueBuilder" will
be executed if any of the terms are true, and otherwise "allFalseBuilder"
will be executed.

The next parameter indicates how many terms there are in the conditional
nest, followed by two parameters for each term: a builder object and
an IlValue. The operations that compute each test should be put into
that term's builder object. The IlValue representing the end result of
the test (zero or non-zero) is also provided.

IlBuilder.cpp shows an example for how to use it. I also added brief
descriptions for these new services in the header and with more detail
in the cpp file.

Issue: #833
Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>