Remove isWarm parm from snippet printing
Warm snippets are deprecated.

* dumpTargetAddressSnippets
* dumpDataSnippets

For the following functions the `isWarm` parameter still remains on the
function signature but the constant 0 has been propagated through and
paths folded as necessary.  This parm could not be eliminated yet
due to downstream project dependencies that need to be resolved.  Once
they are then this optional parm will be removed as well.

* TR_Debug::print(TR::FILE *pOutFile, TR::list<TR::Snippet*> & snippetLista, bool isWarm=0)
* TR_Debug::print(TR::FILE *pOutFile, List<TR::Snippet> & snippetList, bool isWarm=0)

Signed-off-by: Daryl Maier <maier@ca.ibm.com>