Rewrite post- to pre-increment in loop tests
Canonicalizer now looks within loops for inequality tests against the
old value of a variable that is incremented or decremented just prior to
the test, and attempts to rewrite those tests in terms of the new value.
As an example in pseudocode, it attempts to change `if (i++ < n)` into
`if (++i <= n)`.

Particularly for loop tests against induction variables, the optimizer
is better attuned to tests against the new value, because those are
commonly generated for typical for-loops (after canonicalization).

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>