Propagate through relative global constraints in VP
When adding a block constraint, VP (value propagation) may attempt to
propagate it through any relative constraints on the same value number.
For this purpose VP now looks at relative global constraints in addition
to block constraints, which helps it to constrain the value of an
induction variable based on a comparison against its previous value.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>