Remove never-used snippet X86fbits2iSnippet
This snippet was garded as debug() since day one and has never been used or tested.
Hard-coded snippet is generally undesirable as it is difficult to understand or maintain.
In addition, once AVX supported is introduced, hard-coded snippet would not be automatically
converted to AVX and hence bring in penalty. Therefore, removing X86fbits2iSnippet.

Signed-off-by: Victor Ding <dvictor@ca.ibm.com>