Add flag for pending push symbols
The current method to determine if a symbol is a pending push is to
check if it is stored in the method's array of pending pushes or by
checking its CP index. This adds a flag to determine if a symbol is
a pending push, applied to it when it is created.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>