Revert "Add test for CodeMetaDataManager"
This reverts commit 3b689493809ae71df1349369251135fea31b472d. A
downstream product is unable to process this test at the moment, and
there is an infrastructural dependency that prevents the test from being
consumed as is which must first be resolved.

In lieu of having a good fix for this, and because there is no actual
product code dependent on this test for correctness, for now the correct
choice appears to be reverting.

Signed-off-by: Matthew Gaudet <magaudet@ca.ibm.com>