Add infrastructure to support JitBuilder testing
When writing tests for JitBuilder services, some amount of boilerplate
code is required in addition to the test code itself. This commit adds
some services that abstract away some of these parts. The goal of
these services is to make writing tests easier and faster. As a side
effect these test cases should also be easier to read and understand.

Signed-off-by: Leonardo Banderali <leob@linux.vnet.ibm.com>