Remove code related to -Xnoquickstart
-Xnoquickstart is an option introduced as a workaround for micro-benchmarks
that had most/all of the code in a method invoked only once. Thus, even if
the JIT compiled that important method, the transition to the new body
never happened.
This restriction was eliminated with the advent of DLT which allows the
transition from interpreted code to jitted code during the same invocation.
As a consequence -Xnoquickstart is more or less useless now.
This change set deprecates the -Xnoquickstart option (the JIT still recognizes
the option, but does nothing).

For documentation purposes: Xnoquickstart compiles loopy methods from their
first invocation (count=0) and uses hot opt level.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>