Improve reg pressure sim for small 64-bit consts
I came across an example block with a sequence of code using
lconsts to address an array of structs. Unhappy with the
register allocation in the method, I tracked the problem
down to the handling of register pressure for these lconsts:
each one increased register pressure by 1 until the sequence
blew the register budget. The immediate reason for the
increased register pressure is that these lconsts are assumed
to require a register for the duration of the block, even
though the constants are small and generally end up folded
into memory references.

There is an escape hatch for 32-bit integers to say they do
not require a register. This commit adds 32-bit integer
constants masquerading as 64-bit integer constants due to
IL type consistency requirements on 64-bit platforms. It
allows 64-bit integer constants where node->isHighWordZero().

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>