Add UpdateStack API to handle stack moves
In a lot of Virtual Machines the Stack is allocated on the heap.  This
means that the Stack may move.  If the Virtual Machine stack moves we
need an API to update the simulated Operand Stack.

The API needs to update OperandStack_base and _stackTopRegister.  I
updated the operandstacktest to move the stack a few times throughout
the execution and cleaned it up so the stack is properly freed.

Signed-off-by: Charlie Gracie <charlie.gracie@gmail.com>