Use createNamedSymbol for TypeDictionary struct fields
TypeDictionary currently uses a mix of services to create symbols for
the fields of structs, but none of them accept a name parameter so
in JIT logs the symbols all come out as "unknown field". This commit
changes TypeDictionary to use createNamedSymbol instead to pass in
the actual name of the field so it shows up in the logs properly as
"struct.field".

A side effect of this change is that 32-bit integer fields will no
longer be added to the list of generic mutable integer fields which
I believe is acceptable (no other type of field was being added to
this list, and I don't believe 32-bit integer fields are so special
so as to require being in this list).

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>