Avoid printing guard's callee symbol if not specified
Virtual guards may contain the symbol reference for the method they
guard, however, it is not guaranteed. This can be seen when creating
SideEffect guards in VP, where the guard is constructed using the
current block's start node, rather than a call node. However, when
dumping the CHTable, it was assumed that a guard would always
have this symbol reference specified.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>