Inline OMR::Instruction::self()
Because of Issue 915, Instruction.hpp had to arranged in an awkward way
to allow the inlining of self(). Namely, the constructors had to be
declared in the class definition, and then had to be defined outside the
class definition so that the OMRInstruction_inlines.hpp file could be
included at which point TR::Instruction is a complete type and self()
can get inlined properly.

Signed-off-by: Irwin D'Souza <dsouzai@ca.ibm.com>