Assign CCRs before GPRs to avoid blocked registers
This patch makes sure to assign CCRs before GPRs if and
when required when assigning registers for dependencies.
This is to avoid the case where the dependency assigns
and therefore blocks every single available GPR prior to
assigning any CCR. If any CCRs are in the spilled state
and must be reverse spilled we will require an unblocked
GPR to use as a scratch register, since on Power we can't
load/store a physical CCR directly.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>