Crash when -Xjit:timingCumulative option is used
The option -Xjit:timingCumulative is used to print ilGen/optimizer/codegen
times across all compilations. This is performed at shutdown time in a
static method OMR::Compilation::shutdown(TR_FrontEnd * fe).
However, this method internally makes use of the compilation object
retried from local thread storage: TR::comp(). This is invalid because
the shutdown() routine is not executed on a compilation thread.
There is another bug related to TR_SingleTimer code: the function used
to print the time taken uses a static char buffer which makes this routine
not multithreaded safe. I have replaced this timeTakenString() method
with a new one, secondsTaken(), that can be used in a multithreaded
environment.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>